Added command line tools for the materializer(S) and for dl reasoner(S)

+++ src/main/java/orar/materializer/Materializer.java
@@ -21,5 +21,5 @@ public interface Materializer {
 	 */
 	public OrarOntology getOrarOntology();
 
-	public int getReasoningTimeInSeconds();
+	public long getReasoningTimeInSeconds();
 }
